+++ letsStart/src/app.ts
@@ -1,5 +1,5 @@
 import * as express from "express";
-import { Cat, CatType } from "./app.models";
+import catsRouter from "./cats/cats.route";
 
 const app: express.Express = express();
 
@@ -13,63 +13,7 @@ app.use((req, res, next) => {
 //* json middleware
 app.use(express.json());
 
-//* READ 고양이 전체 데이터 조회
-app.get("/cats", (req: express.Request, res: express.Response) => {
-  try {
-    const cats = Cat;
-    // throw new Error("db connect error");
-    res.send({
-      success: true,
-      data: {
-        cats,
-      },
-    });
-  } catch (error: any) {
-    res.status(400).send({
-      success: false,
-      error: error.message,
-    });
-  }
-});
-
-//* READ 특정 고양이 데이터 조회
-app.get("/cats/:id", (req: express.Request, res: express.Response) => {
-  try {
-    const params = req.params;
-    const cats = Cat.find((cat) => {
-      return cat.id === params.id;
-    });
-    res.send({
-      success: true,
-      data: {
-        cats,
-      },
-    });
-  } catch (error: any) {
-    res.status(400).send({
-      success: false,
-      error: error.message,
-    });
-  }
-});
-
-//* CREATE 새로운 고양이 추가 api
-app.post("/cats", (req: express.Request, res: express.Response) => {
-  try {
-    const data = req.body;
-    console.log(data);
-    Cat.push(data); // create
-    res.status(200).send({
-      success: true,
-      data: { data },
-    });
-  } catch (error: any) {
-    res.status(400).send({
-      success: false,
-      error: error.message,
-    });
-  }
-});
+app.use(catsRouter);
 
 // * 404 middlewart
 app.use((req, res, next) => {
